refactor: move api to engine package and create snake interface
