+++ tests/entity/test_apple.py
@@ -1,5 +1,5 @@
 import unittest
-from api.entity.apple import Apple, AppleType
+from engine.entity.apple import Apple, AppleType
 
 
 class TestApple(unittest.TestCase):
